python/shared/pyproject.toml
[project]
name = "data-go-kr-shared"
version = "0.1.0"
description = "공공데이터 포털 MCP 서버 공통 유틸리티"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_go_kr_shared"]
